labapp/requirements.txt
alembic==1.0.2
asn1crypto==0.24.0
Babel==2.6.0
blinker==1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.0.1
cryptography==2.4.2
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
openapi-spec-validator==0.2.4
passlib==1.7.1
pkg-resources==0.0.0
pycparser==2.19
PyJWT==1.6.4
PyMySQL==0.9.2
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.7
PyYAML==3.13
requests==2.20.1
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.13
swagger-ui-bundle==0.0.2
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
